began setting up router, exported router functions to new file for reusability, began setting up saga
